from ("%u") to ("%d",(int)) for size_t
svn path=/trunk/iCub/tutorials/; revision=15995

+++ src/tutorial_gaze_interface.cpp
@@ -250,8 +250,8 @@ public:
             // (i.e. torso and head angles zeroed). 
             igaze->getAngles(ang);            
 
-            fprintf(stdout,"Storing POI #%u ... %s [deg]\n",
-                    poiList.size(),ang.toString().c_str());
+            fprintf(stdout,"Storing POI #%d ... %s [deg]\n",
+                    (int)poiList.size(),ang.toString().c_str());
 
             poiList.push_back(ang);
